src/test/java/com/octo/persistence/specification/filter/TokenPredicateFilterTest.java
package com.octo.persistence.specification.filter;

import com.octo.helper.MockHelper;
import com.octo.persistence.model.Project;
import com.octo.persistence.specification.PredicateOperator;
import org.junit.jupiter.api.Tag;
import org.junit.jupiter.api.Test;
import org.junit.jupiter.api.extension.ExtendWith;
import org.springframework.test.context.junit.jupiter.SpringExtension;

import javax.persistence.EntityManager;
import javax.persistence.criteria.CriteriaBuilder;
import javax.persistence.criteria.CriteriaQuery;
import javax.persistence.criteria.Root;

import static org.junit.jupiter.api.Assertions.*;

@ExtendWith(SpringExtension.class)
@Tag("unit")
class TokenPredicateFilterTest extends MockHelper {

    @Test
    void constructorTest() {
        TokenPredicateFilter filter = new TokenPredicateFilter(null, null);

        assertNull(filter.getName());
        assertEquals(0, filter.getValues().length);
    }

    @Test
    void testGetPredicate() {
        EntityManager entityManager = mockEntityManager(Project.class);
        final CriteriaBuilder builder = entityManager.getCriteriaBuilder();
        final CriteriaQuery<Project> query = builder.createQuery(Project.class);
        final Root<Project> root = query.from(Project.class);

        TokenPredicateFilter filter = new TokenPredicateFilter("token", "test");
        assertTrue(filter.extract());
        assertNotNull(filter.getPredicate(builder, root, null));
        assertEquals(PredicateOperator.EQUALS, filter.getOperator(0));

        filter = new TokenPredicateFilter("token", "not_test");
        assertTrue(filter.extract());
        assertNotNull(filter.getPredicate(builder, root, null));
        assertEquals(PredicateOperator.EQUALS, filter.getOperator(0));
    }
}
